requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
argon2-cffi=20.1.0=py38h2bbff1b_1
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
blosc=1.21.0=h19a0ad4_0
brotli=1.0.9=ha925a31_2
bzip2=1.0.8=he774522_0
ca-certificates=2021.10.8=h5b45459_0
certifi=2021.10.8=py38haa244fe_1
cffi=1.15.0=py38h2bbff1b_0
cfitsio=3.470=he774522_6
charls=2.2.0=h6c2663c_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py38he774522_0
dask-core=2021.10.0=pyhd3eb1b0_0
debugpy=1.5.1=py38hd77b12b_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py38_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
fsspec=2021.10.1=pyhd3eb1b0_0
giflib=5.2.1=h62dcd97_0
icc_rt=2019.0.0=h0cc432a_1
imagecodecs=2021.8.26=py38ha1f97ea_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=4.8.2=py38haa95532_0
importlib_metadata=4.8.2=hd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
ipydatawidgets=4.2.0=pyhd3deb0d_0
ipykernel=6.4.1=py38haa95532_1
ipython=7.29.0=py38hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.0=py38haa95532_1
jinja2=3.0.2=pyhd3eb1b0_0
jpeg=9d=h2bbff1b_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.0.6=pyhd3eb1b0_0
jupyter_core=4.9.1=py38haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.1=py38hd77b12b_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libaec=1.0.4=h33f27b4_1
libdeflate=1.8=h2bbff1b_5
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libwebp=1.2.0=h2bbff1b_0
libzopfli=1.0.3=ha925a31_0
locket=0.2.1=py38haa95532_1
lz4-c=1.9.3=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.0.1=py38h2bbff1b_0
matplotlib-base=3.5.0=py38h6214cd6_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
meshplot=0.4.0=pyhd3deb0d_2
mistune=0.8.4=py38he774522_1000
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=py_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py38haa95532_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.6.3=pyhd3eb1b0_0
notebook=6.1.4=py38_0
numpy=1.21.2=py38hfca59bb_0
numpy-base=1.21.2=py38h0829f74_0
olefile=0.46=pyhd3eb1b0_0
openjpeg=2.4.0=h4fc8c34_0
openssl=1.1.1l=h8ffe710_0
packaging=21.3=pyhd3eb1b0_0
pandocfilters=1.4.3=py38haa95532_1
parso=0.8.2=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py38hd45dc43_0
pip=21.2.2=py38haa95532_0
prometheus_client=0.12.0=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.10.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyrsistent=0.18.0=py38h196d8e1_0
python=3.8.8=hdbf39b2_5
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pythreejs=2.3.0=pyhd8ed1ab_0
pywavelets=1.1.1=py38he774522_2
pywin32=302=py38h827c3e9_1
pywinpty=0.5.7=py38_0
pyyaml=6.0=py38h2bbff1b_1
pyzmq=22.3.0=py38hd77b12b_2
scikit-image=0.18.3=py38hf11a4ad_0
scipy=1.7.1=py38hbe87c03_2
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py38haa95532_0
six=1.16.0=pyhd3eb1b0_0
snappy=1.1.8=h33f27b4_0
sqlite=3.36.0=h2bbff1b_0
terminado=0.9.4=py38haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.11=h2bbff1b_0
toolz=0.11.2=pyhd3eb1b0_0
tornado=6.1=py38h2bbff1b_0
traitlets=5.1.1=pyhd3eb1b0_0
traittypes=0.2.1=pyh9f0ad1d_2
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.1=py38_0
wincertstore=0.2=py38haa95532_2
winpty=0.4.3=4
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
zfp=0.5.5=hd77b12b_6
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h8cc25b3_4
zstd=1.4.9=h19a0ad4_0
